Changes sunday

+++ DAL Interface/Repository/IGenericRepository.cs
@@ -9,7 +9,7 @@ namespace DAL_Interface.Repository
     public interface IGenericRepository<TEntity> where TEntity : class
     {
         void Create(TEntity item);
-        TEntity FindById(int id);
+        TEntity FindById(long id);
         IEnumerable<TEntity> Get();
         IEnumerable<TEntity> Get(Func<TEntity, bool> predicate);
         void Remove(TEntity item);
